# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервера для расчета фотоэффекта",
        "title": "RIP Golab Photoeffect API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/calculation-items": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Корзина"
                ],
                "summary": "Обновление элемента в корзине",
                "parameters": [
                    {
                        "description": "Новые параметры",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.M2MInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Корзина"
                ],
                "summary": "Добавление излучения в корзину (черновик)",
                "parameters": [
                    {
                        "description": "Параметры эксперимента",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.M2MInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Корзина"
                ],
                "summary": "Удаление элемента из корзины",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Излучения",
                        "name": "radiation_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешно удалено"
                    }
                }
            }
        },
        "/api/calculations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявки"
                ],
                "summary": "Список всех сформированных заявок (с фильтрами)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата от (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата до (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/api/calculations/draft-summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявки"
                ],
                "summary": "Получение иконки корзины (сводка черновика)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/calculations/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Заявки"
                ],
                "summary": "Получение заявки по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP_Golab_internal_app_ds.RadiationCalculation"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Заявки"
                ],
                "summary": "Изменение описания заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Описание",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Заявки"
                ],
                "summary": "Логическое удаление заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешно"
                    }
                }
            }
        },
        "/api/calculations/{id}/complete": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только Профессору",
                "tags": [
                    "Заявки"
                ],
                "summary": "Завершение или отклонение заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Action: accept / reject",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/calculations/{id}/form": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Переводит заявку из draft в сформирован и считает формулы",
                "tags": [
                    "Заявки"
                ],
                "summary": "Сформировать заявку (запуск расчетов)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/radiations": {
            "get": {
                "description": "Возвращает список всех излучений из каталога",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Услуги"
                ],
                "summary": "Получение всех услуг",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Поиск по названию",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RIP_Golab_internal_app_ds.RadiationRange"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только Профессору",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Услуги"
                ],
                "summary": "Добавление новой услуги",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Описание",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Изображение",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Видео",
                        "name": "video",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/RIP_Golab_internal_app_ds.RadiationRange"
                        }
                    }
                }
            }
        },
        "/api/radiations/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Услуги"
                ],
                "summary": "Получение услуги по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Излучения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP_Golab_internal_app_ds.RadiationRange"
                        }
                    }
                }
            }
        },
        "/api/users/login": {
            "post": {
                "description": "Выдает JWT токен в случае успешной авторизации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пользователи"
                ],
                "summary": "Авторизация (Login)",
                "parameters": [
                    {
                        "description": "Креды (login, password)",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/users/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Помещает переданный JWT в Blacklist Redis'а",
                "tags": [
                    "Пользователи"
                ],
                "summary": "Выход (Logout)",
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "RIP_Golab_internal_app_ds.RadiationRange": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_delete": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
...
        "internal_app_handler.M2MInput": {
            "type": "object",
            "required": [
                "radiation_id"
            ],
            "properties": {
                "area": {
                    "type": "number"
                },
                "efficiency": {
                    "type": "number"
                },
                "frequency": {
                    "type": "number"
                },
                "radiation_id": {
                    "type": "integer"
                },
                "work_function": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}